requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot
psycopg2-binary
feedparser
pyowm
googletrans
wikipedia
kbbi
Pillow
alphabet_detector
urbandict
tldextract
coffeehouse
covid
